Module dkg_gadget::gossip_engine

Webb Custom DKG Gossip Engine.

Structs

GossipHandler

Handler for gossiping messages. Call GossipHandler::run to start the processing.

GossipHandlerController

Controls the behaviour of a GossipHandler it is connected to.

HandshakeMessage

A Handshake message that is sent when a peer connects to us, to verify that the peer Id (which is the sender) is the owner of the authority id. This is used to prevent a malicious peer from impersonating another authority. Notes:

NetworkGossipEngineBuilder

Enums

DKGNetworkMessage

Traits

GossipEngineIface

A GossipEngine that can be used to send DKG messages.