Trait dkg_gadget::gossip_engine::GossipEngineIface

pub trait GossipEngineIface: Send + Sync + 'static {
    type Clock: AtLeast32BitUnsigned + Send + Sync + 'static;

    fn send(
        &self,
        recipient: PeerId,
        message: SignedDKGMessage<AuthorityId>
    ) -> Result<(), DKGError>;
    fn gossip(
        &self,
        message: SignedDKGMessage<AuthorityId>
    ) -> Result<(), DKGError>;
    fn message_available_notification(
        &self
    ) -> Pin<Box<dyn Stream<Item = ()> + Send>>;
    fn peek_last_message(&self) -> Option<SignedDKGMessage<AuthorityId>>;
    fn acknowledge_last_message(&self);
    fn clear_queue(&self);
    fn local_peer_id(&self) -> PeerId;
    fn logger(&self) -> &DebugLogger;
}

A GossipEngine that can be used to send DKG messages.

in the core it is very simple, just two methods:

send which will send a DKG message to a specific peer.

gossip which will send a DKG message to all peers.

stream which will return a stream of DKG messages.

Required Associated Types

type Clock: AtLeast32BitUnsigned + Send + Sync + 'static

Required Methods

fn send(
&self,
recipient: PeerId,
message: SignedDKGMessage<AuthorityId>
) -> Result<(), DKGError>

Send a DKG message to a specific peer.

fn gossip(&self, message: SignedDKGMessage<AuthorityId>) -> Result<(), DKGError>

Send a DKG message to all peers.

fn message_available_notification(
&self
) -> Pin<Box<dyn Stream<Item = ()> + Send>>

A stream that sends messages when they are ready to be polled from the message queue.

fn peek_last_message(&self) -> Option<SignedDKGMessage<AuthorityId>>

Peek the front of the message queue.

Note that this will not remove the message from the queue, it will only return it. For removing the message from the queue, use acknowledge_last_message.

Returns None if there are no messages in the queue.

fn acknowledge_last_message(&self)

Acknowledge the last message (the front of the queue) and mark it as processed, then removes it from the queue.

fn clear_queue(&self)

Clears the Message Queue.

fn local_peer_id(&self) -> PeerId

fn logger(&self) -> &DebugLogger