Module dkg_gadget::worker

Structs

DKGWorker

A DKG worker plays the DKG protocol

TestBundle

Used only for tests

WorkerParams

Constants

ENGINE_ID

MAX_KEYGEN_RETRIES

MAX_SIGNING_SETS

MAX_SUBMISSION_DELAY

MAX_UNSIGNED_PROPOSALS_PER_SIGNING_SET

PROPOSAL_HASH_LIFETIME

How many blocks to keep the proposal hash in out local cache.

STORAGE_SET_RETRY_NUM

Traits

HasLatestHeader

KeystoreExt

Extension trait for any type that contains a keystore

Type Definitions

AggregatedMisbehaviourReportStore

Shared