Crate dkg_gadget

Re-exports

pub use constants::DKG_KEYGEN_PROTOCOL_NAME;

pub use constants::DKG_SIGNING_PROTOCOL_NAME;

pub use keystore::DKGKeystore;

Modules

async_protocols

constants

db

deadlock_detection

debug_logger

dkg_modules

gossip_engine

Webb Custom DKG Gossip Engine.

gossip_messages

keyring

Stores keypairs for DKG

keystore

signing_manager

storage

worker

Macros

metric_inc

metric_set

Structs

DKGParams

DKG gadget initialization parameters.

Enums

RoundsEventType

Traits

Client

A convenience DKG client trait that defines all the type bounds a DKG client has to satisfy. Ideally that should actually be a trait alias. Unfortunately as of today, Rust does not allow a type alias to be used as a trait bound. Tracking issue is https://github.com/rust-lang/rust/issues/41517.

Functions

dkg_peers_set_config

Returns the configuration value to put in [sc_network::config::NetworkConfiguration::extra_sets].

start_dkg_gadget

Start the DKG gadget.